+++ Makefile
@@ -16,7 +16,21 @@ override LIBS += $(shell pkg-config --libs glib-2.0)
 CFLAGS ?= -std=c99 -Os -Wall -Wextra
 override CFLAGS += $(shell pkg-config --cflags glib-2.0) -DVERSION=\"$(VERSION)\" -DGIT_COMMIT=\"$(GIT_COMMIT)\"
 
-bin/conmon: src/conmon.o src/cmsg.o | bin
+# Conditionally compile journald logging code if the libraries can be found
+# if they can be found, set USE_JOURNALD macro for use in conmon code.
+#
+# "pkg-config --exists" will error if the package doesn't exist. Make can only compare
+# output of commands, so the echo commands are to allow pkg-config to error out, make to catch it,
+# and allow the compilation to complete.
+ifeq ($(shell pkg-config --exists libsystemd-journal && echo "0" || echo "1"), 0)
+	override LIBS += $(shell pkg-config --libs libsystemd-journal)
+	override CFLAGS += $(shell pkg-config --cflags libsystemd-journal) -D USE_JOURNALD=0
+else ifeq ($(shell pkg-config --exists libsystemd && echo "0" || echo "1"), 0)
+	override LIBS += $(shell pkg-config --libs libsystemd)
+	override CFLAGS += $(shell pkg-config --cflags libsystemd) -D USE_JOURNALD=0
+endif
+
+bin/conmon: src/conmon.o src/cmsg.o src/ctr_logging.o src/utils.o | bin
 	$(CC) $(LDFLAGS) -o $@ $^ $(LIBS)
 
 %.o: %.c
@@ -24,7 +38,11 @@ bin/conmon: src/conmon.o src/cmsg.o | bin
 
 src/cmsg.o: src/cmsg.c src/cmsg.h
 
-src/conmon.o: src/conmon.c src/cmsg.h src/config.h
+src/utils.o: src/utils.c src/utils.h
+
+src/ctr_logging.o: src/ctr_logging.c src/ctr_logging.h src/utils.h
+
+src/conmon.o: src/conmon.c src/cmsg.h src/config.h src/utils.h src/ctr_logging.h
 
 bin:
 	mkdir -p bin
